+++ src/app/dashboard/plus-badge/plus-badge.component.ts
@@ -1,0 +1,8 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'plus-badge',
+  templateUrl: './plus-badge.component.html',
+  styleUrls: ['./plus-badge.component.css']
+})
+export class PlusBadgeComponent {}