+++ src/app/guards/dashboard-auth.guard.ts
@@ -10,11 +10,7 @@ export class DashboardAuthGuard implements CanActivate {
   constructor(private userService: UserService) {}
 
   async canActivate() {
-    if (!this.userService.user)
-      await this.userService.updateUser();
-    if (!this.userService.savedUser)
-      await this.userService.updateSavedUser(); 
-        
+    await this.userService.init();
     return Boolean(this.userService.user);
   }  
 }