+++ src/app/logout/logout.component.ts
@@ -18,7 +18,6 @@ export class LogoutComponent implements OnInit {
     localStorage.removeItem('key');
     
     await this.userService.updateUser();
-    await this.userService.updateSavedUser();
     await this.guildService.updateGuilds();
 
     this.router.navigate(['/']);