+++ src/environments/environment.ts
@@ -9,6 +9,7 @@ export const environment = {
   githubURL: 'https://github.com/theADAMJR/2PG',
   stripePublicKey: 'pk_test_rjQKo5c9pcHvDFjFe1mfdKib00Es3eJ831',
   version: 'v0.2.1b',
+  botOwnerId: '218459216145285121',
   test: {
     guildId: '685862664223850497'
   }